Make windows ignore restoring non-existent state/geometry
This fixes a bug related to the main window opening in the top left
corner of the screen instead of the center when starting up LibrePlan
for the first time.

This bug originates from the main window immediately setting its own
state/geometry, resulting in them being put in the top left corner of
the screen.

This fix instead makes `Config` ignore any restoration requests unless
the window state/geometry is in the config table.