Add a version requirement for PyInstaller
With PyInstaller versions 6.0.0 and above, the `resources` directory was
put under a directory the main executable didn't expect, resulting in
those resources not being loaded.

The fix is to put the `_internal` directory in the directory as the main
executable, as versions of PyInstaller prior to 6.0.0 did.

As such, the PyInstaller's version is now restricted to 6.*, and by
extension, all other requirements that expected to have breaking changes
will have their versions restricted.