Add settings for when backups are created
This commit adds two settings which determine when backups are made,
either on exit and/or upon Plan completing.

This change was made because a user might keep LibrePlan open constantly,
but would still like to have automatic backups made periodically
(ideally, daily, which is why one of the settings makes backups when
Plan is completed). However, making backups when Plan is completed might
be redundant if the user regularly exits LibrePlan afterwards, so we
make it an optional feature.