Make backup file names more unique
Before, if multiple backups were being made in the same second (such as
when the backup system was being tested), the new backup files would
overwrite the old ones. This commit fixes this by appending a number to
the end of the file name if the file already exists.

This commit also changes the backup file name filter used when
performing old backup file auto-cleanup to include anything after the
date and time, so the file suffix can be something else in the future.